chore: split up theme to provider and hooks
