+++ src/hooks/theme/theme-hook.tsx
@@ -1,0 +1,10 @@
+import { ThemeProviderContext } from '@/providers/theme/theme-provider';
+import { useContext } from 'react';
+
+export const useTheme = () => {
+  const context = useContext(ThemeProviderContext);
+
+  if (context === undefined) throw new Error('useTheme must be used within a ThemeProvider');
+
+  return context;
+};
